requirements.txt
asgiref==3.8.1
Django==5.1.3
django-ckeditor==6.7.1
django-classy-tags==4.1.0
django-formtools==2.5.1
django-js-asset==2.2.0
django-render-partial==0.4
djangocms-admin-style==3.3.1
mysqlclient==2.2.5
packaging==24.2
pillow==11.0.0
PyMySQL==1.1.1
python-decouple==3.8
setuptools==75.5.0
sqlparse==0.5.1
tzdata==2024.2
